Generate the CLO file for use with the UNIBE tex template

The CLO file contains the information identifying the author, project and institution (although we do not modify that). It is used by the UNIBE CLS file to create the parts of the template.

use_ubreportclo(
  projnum = "xxx",
  projname = "Project YYY",
  reporttype = "Type of report",
  version = "Version",
  sign = "Author name",
  email = "author.name@ctu.unibe.ch",
  job = "Senior Statistician",
  open = FALSE,
  dir = ".",
  ...
)

Arguments

projnum

Project number to appear in report

projname

Project name to appear in report

reporttype

The main title of the report

version

Version number or similar

sign

Your name as it should appear

email

your email address

job

your job title

open

open the file or not

dir

directory in which to create the file

...

other options passed to use_template

Value

writes the UNIBE template clo file

Examples

# use_ubreportclo(# Personal info
#                 sign = "Alan",
#                 email = "alan.haynes@ctu.unibe.ch",
#                 job = "Senior Statistician",
#                 # Project info
#                 projnum = "1234",
#                 projname = "Project X",
#                 # Report info
#                 version = Sys.Date(),
#                 reporttype = "Recruitment report"
#                 )